+++ R/receptiviti.R
@@ -18,7 +18,8 @@
 #' @param files A list of file paths, as alternate entry to \code{text}.
 #' @param dir A directory to search for files in, as alternate entry to \code{text}.
 #' @param file_type File extension to search for, if \code{text} is the path to a directory containing files to be read in.
-#' @param encoding Encoding of a txt file to be read in; passed to \code{\link{readLines}}.
+#' @param encoding Encoding of file(s) to be read in. If not specified, this will be detected, which can fail,
+#' resulting in mis-encoded characters; for best (and fasted) results, specify encoding.
 #' @param return_text Logical; if \code{TRUE}, \code{text} is included as the first column of the result.
 #' @param api_args A list of additional arguments to pass to the API (e.g., \code{list(sallee_mode = "sparse")}). Defaults to the
 #' \code{receptiviti.api_args} option.
@@ -164,12 +165,13 @@
 #' @importFrom parallel detectCores makeCluster clusterExport parLapplyLB parLapply stopCluster
 #' @importFrom future.apply future_lapply
 #' @importFrom progressr progressor
-#' @importFrom arrow read_csv_arrow write_csv_arrow schema string write_dataset open_dataset
+#' @importFrom arrow read_csv_arrow write_csv_arrow schema string write_dataset open_dataset CsvReadOptions
 #' @importFrom dplyr filter compute collect select all_of
+#' @importFrom stringi stri_enc_detect
 #' @export
 
 receptiviti <- function(text, output = NULL, id = NULL, text_column = NULL, id_column = NULL, files = NULL, dir = NULL,
-                        file_type = "txt", encoding = "unknown", return_text = FALSE, api_args = getOption("receptiviti.api_args", list()),
+                        file_type = "txt", encoding = NULL, return_text = FALSE, api_args = getOption("receptiviti.api_args", list()),
                         frameworks = getOption("receptiviti.frameworks", "all"), framework_prefix = TRUE, as_list = FALSE,
                         bundle_size = 1000, bundle_byte_limit = 75e5, collapse_lines = FALSE, retry_limit = 50, clear_cache = FALSE,
                         clear_scratch_cache = TRUE, request_cache = TRUE, cores = detectCores() - 1, use_future = FALSE,
@@ -217,6 +219,13 @@ receptiviti <- function(text, output = NULL, id = NULL, text_column = NULL, id_c
     if (!all(file.exists(text))) stop("not all of the files in text exist", call. = FALSE)
   }
   read_in <- FALSE
+  handle_encoding <- function(file) {
+    if (is.null(encoding)) {
+      unlist(stringi::stri_enc_detect(readBin(file, "raw", 200))[[1]])[[1]]
+    } else {
+      encoding
+    }
+  }
   if (text_as_dir || text_as_paths || (is.character(text) && !anyNA(text) && all(nchar(text) < 500))) {
     if (text_as_dir || length(text) == 1 && dir.exists(text)) {
       if (verbose) message("reading in texts from directory: ", text, " (", round(proc.time()[[3]] - st, 4), ")")
@@ -232,10 +241,9 @@ receptiviti <- function(text, output = NULL, id = NULL, text_column = NULL, id_c
           if (is.null(text_column)) stop("text appears to point to csv files, but text_column was not specified", call. = FALSE)
           read_in <- TRUE
           text <- unlist(lapply(text, function(f) {
-            d <- tryCatch(
-              read_csv_arrow(f, col_select = c(text_column, id_column)),
-              error = function(e) NULL
-            )
+            d <- tryCatch(read_csv_arrow(f, read_options = CsvReadOptions$create(
+              encoding = handle_encoding(f)
+            ), col_select = c(text_column, id_column)), error = function(e) NULL)
             if (is.null(d)) stop("failed to read in file ", f, call. = FALSE)
             if (!is.null(id_column) && id_column %in% colnames(d)) {
               structure(d[, text_column, drop = TRUE], names = d[, id_column, drop = TRUE])
@@ -245,13 +253,20 @@ receptiviti <- function(text, output = NULL, id = NULL, text_column = NULL, id_c
           }))
         } else {
           text <- unlist(lapply(text, function(f) {
-            d <- readLines(f, warn = FALSE, encoding = encoding, skipNul = TRUE)
-            d[d != ""]
+            tryCatch(
+              {
+                con <- file(f, encoding = handle_encoding(f))
+                on.exit(close(con))
+                d <- readLines(con, warn = FALSE, skipNul = TRUE)
+                d[d != ""]
+              },
+              error = function(e) stop("failed to read in file ", f, call. = FALSE)
+            )
           }))
         }
         id <- names(text)
       }
-    } else if (length(text) == 1 && dirname(text) != "." && dir.exists(dirname(text))) {
+    } else if (length(text) == 1 && dirname(text) != "." && dir.exists(dirname(dirname(text)))) {
       stop("text appears to be a directory, but it does not exist")
     }
     if (text_as_paths && missing(id)) {
@@ -496,9 +511,25 @@ receptiviti <- function(text, output = NULL, id = NULL, text_column = NULL, id_c
     if (text_as_paths) {
       if (all(grepl("\\.csv", text, TRUE))) {
         if (is.null(text_column)) stop("files appear to be csv, but no text_column was specified", call. = FALSE)
-        text <- vapply(text, function(f) paste(arrow::read_csv_arrow(f, col_select = all_of(text_column))[[1]], collapse = " "), "")
+        text <- vapply(text, function(f) {
+          tryCatch(
+            paste(arrow::read_csv_arrow(f, read_options = arrow::CsvReadOptions$create(
+              encoding = handle_encoding(f)
+            ), col_select = all_of(text_column))[[1]], collapse = " "),
+            error = function(e) stop("failed to read in file ", f, call. = FALSE)
+          )
+        }, "")
       } else {
-        text <- vapply(text, function(f) paste(readLines(f, warn = FALSE, encoding = encoding, skipNul = TRUE), collapse = " "), "")
+        text <- vapply(text, function(f) {
+          tryCatch(
+            {
+              con <- file(f, encoding = handle_encoding(f))
+              on.exit(close(con))
+              paste(readLines(con, warn = FALSE, skipNul = TRUE), collapse = " ")
+            },
+            error = function(e) stop("failed to read in file ", f, call. = FALSE)
+          )
+        }, "")
       }
     }
     bundle$hashes <- paste0(vapply(paste0(args_hash, text), digest::digest, "", serialize = FALSE))
@@ -570,7 +601,7 @@ receptiviti <- function(text, output = NULL, id = NULL, text_column = NULL, id_c
     for (name in c(
       "doprocess", "request", "process", "text_column", "prog", "make_request", "check_cache", "full_url",
       "temp", "use_future", "cores", "bundles", "cache_format", "request_cache", "auth",
-      "text_as_paths", "retry_limit", "api_args", "args_hash", "encoding"
+      "text_as_paths", "retry_limit", "api_args", "args_hash", "encoding", "handle_encoding"
     )) {
       call_env[[name]] <- get(name)
     }